Aesthetic Changes
This commit changes the background to something that better fits our aesthetic.

The previous commits by me were changing the layout of the nav bar so that it's now vertical and to the side.

I also moved the posts to the center and adjusted the color scheme to be more UCLA-like.